added a custom confirmation modal to replace the default JavaScript one

+++ client/src/app/messages/messages.component.ts
@@ -2,6 +2,7 @@ import { Component, OnInit } from '@angular/core';
 import { PaginationComponent } from 'ngx-bootstrap/pagination';
 import { Message } from '../_models/message';
 import { Pagination } from '../_models/pagination';
+import { ConfirmService } from '../_services/confirm.service';
 import { MessageService } from '../_services/message.service';
 
 @Component(
@@ -19,7 +20,8 @@ export class MessagesComponent implements OnInit {
   pageSize = 5;
   loading: boolean = false;
 
-  constructor(private messageService: MessageService) { }
+  constructor(private messageService: MessageService,
+    private confirmService: ConfirmService) { }
 
   ngOnInit(): void
   {
@@ -40,9 +42,17 @@ export class MessagesComponent implements OnInit {
 
   deleteMessage(id: number)
   {
-    this.messageService.deleteMessage(id).subscribe(() => {
-      this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
-    });
+    this.confirmService.confirm('Delete Message',
+     'Are you sure? This cannot be undone?').subscribe((result) =>
+     {
+      if(result)
+      {
+        this.messageService.deleteMessage(id).subscribe(() =>
+        {
+          this.messages.splice(this.messages.findIndex(m => m.id === id), 1);
+        });
+      }
+     })
   }
